requirements.txt
fastapi
uvicorn
youtube-transcript-api==0.6.2
groq
python-dotenv